backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Cors==4.0.0
Pillow==10.4.0
numpy<2
matplotlib<3.9
torch
ultralytics
